fix(runtime) removes context factory indirection
This patch removes `root_context_factory_map`, used to store
the context factory associated with the name of the filter as declared
in Envoy configuration file.

During `proxy_on_context_create` execution the property `plugin_root_id`
was retrieved and used to recover the context factory from the map.
That workflow however, doesn't work with ngx_wasm_module as it returns
a numeric value for `plugin_root_id`.

Since it's not clear the necessity of a filter to store multiple root
context factories, the factory map is replaced with a variable that
directly points to the context factory passed to either
`registerRootContext` or `registerRootContextFactory`.

`registerRootContextFactory` is introduced as a simplified version of
`registerRootContext` which doesn't receive a `name` parameter -- now
irrelevant. `registerRootContext` is kept in order to preserve
compatibility with existing filters; it's the paramter `name` now is
simply ignored.